HTTP/FTP file transfers.
Servers can set sv_dlURL to a HTTP/FTP pk3 repository.
Clients can set a secondary cl_dlURL.
If cl_allowDownload is set to 2, cl_dlURL will be used instead of a valid sv_dlURL.
This is a slightly modified version of that contributed to ioquake3 by Tony J. White and Hobbes.
https://bugzilla.icculus.org/show_bug.cgi?id=2661#c20


git-svn-id: https://svn.mode-x.org/cnq3/trunk@880 df7340c0-8d17-0410-b57c-d8c7d96bbbdc